Refactor di Game e Master
