--- server/src/main/scala/it/unibo/scalapacman/server/ScalaGreeter.scala
@@ -1,13 +1,0 @@
-package it.unibo.scalapacman.server
-
-trait ScalaGreeter {
-  def sayHello: String
-}
-
-object ScalaGreeter {
-  def apply(): ScalaGreeter = new ScalaGreeterImpl
-}
-
-class ScalaGreeterImpl extends ScalaGreeter {
-  override def sayHello: String = "Hello, World!"
-}
